Compile with QPE configuration.     - no properties     - no transformations     - fix up some QT_NO_*'s     - QVariant now required for signals and slots Reduce memory consumption.     - move big QText* [virtual] inlines to a .cpp file
[git-p4: depot-paths = "//depot/qt/main/": change = 37836]

+++ src/kernel/qwindowsystem_qws.h
@@ -161,7 +161,7 @@ public:
 				bool autoRepeat);
 
     typedef struct KeyMap {
-	int  key_code;
+	ushort key_code;
 	ushort unicode;
 	ushort shift_unicode;
 	ushort ctrl_unicode;
